fix: preserve split modules after upstream merge
